Add component storage and a query system
This commit adds two methods for fetching a component, mutably or immutably,
for an entity. It also adds a simple way to query and iterate over the Ecs
using the newly added QueryIter iterator.

Currently it only supports querying for 2 components, this will be generalized
in a later commit. Queries only fetch immutable references to the components.