Add tuple impls for Query trait and ExactSizeIterator impl
I added a tuple_query_impl macro that generates impl of the Query traits for
tuples of size 1 to 26. In order to count entities matching a query, I added
an implementation of ExactSizeIterator for QueryIter.

The macro and the iterator implementation can surely be improved. But that's a first step :^)