Add remove_entity method to Ecs
This allows to remove an entity from the Ecs. When an
entity is removed, all its components are set to None in
the component storage and the entity's id is stored in a
free list. When creating a new entity, we fetch an id from
the free list or we allocate more component storage space.